Moved spqlios_alias and rng to backend
