+++ include/core/glwe/bivariate_polynomial.h
@@ -2,6 +2,7 @@
 #define BIVARIATE_POLYNOMIAL_H
 
 #include "glwe_params.h"
+#include "maths_structures.h"
 #include "spqlios_alias.h"
 
 // BIV POLY PART (begin)
