+++ src/common/CMakeLists.txt
@@ -6,8 +6,6 @@
 set(COMMON_SRCS
     logger.c
     maths_structures.c
-    spqlios_alias.c
-    rng.c
     utils.c
 )
 
@@ -18,12 +16,10 @@ add_library(PVDA_COMMON SHARED ${COMMON_SRCS})
 set_target_properties(PVDA_COMMON PROPERTIES OUTPUT_NAME pvdacommon)
 
 # Link dependencies
-target_link_libraries(PVDA_COMMON PRIVATE libspqlios)
+target_link_libraries(PVDA_COMMON PRIVATE)
 
 # Include directories
 target_include_directories(PVDA_COMMON
     PUBLIC
         ${PROJECT_SOURCE_DIR}/include/common                 # Public headers for consumers
-    PRIVATE
-        ${PROJECT_SOURCE_DIR}/build/_deps/spqlios-src/spqlios/arithmetic/  # Internal headers only
 )
