+++ src/core/CMakeLists.txt
@@ -8,5 +8,5 @@ add_subdirectory(ggsw)
 
 # Create CORE interface library and link GLWE
 add_library(PVDA_CORE INTERFACE)
-target_link_libraries(PVDA_CORE INTERFACE PVDA_GLWE PVDA_GGSW PVDA_COMMON)
+target_link_libraries(PVDA_CORE INTERFACE PVDA_GLWE PVDA_GGSW PVDA_COMMON PVDA_BACKEND)
