+++ tests/unit/core/ggsw/test_glwegadget_half_product.c
@@ -2,14 +2,11 @@
 #include <criterion/new/assert.h>
 #include <float.h>
 #include <stdint.h>
-#include <stdio.h>
 
 #include "bivariate_polynomial.h"
-#include "core/ggsw/ggsw_arithmetic.h"
 #include "core/glwe/glwe_arithmetic.h"
 #include "core/glwe/glwe_ciphertext.h"
 #include "core/glwe/glwe_transform_key.h"
-#include "ggsw_params.h"
 #include "glwe_key.h"
 #include "glwe_params.h"
 #include "glwegadget_arithmetic.h"
