+++ tests/unit/core/glwe/test_univariate_polynomial.c
@@ -4,9 +4,8 @@
 #include <math.h>
 #include <stdint.h>
 
-#include "common/rng.h"
-#include "common/spqlios_alias.h"
 #include "glwe_params.h"
+#include "rng.h"
 #include "test_utils.h"
 #include "univariate_polynomial.h"
